sync-github-releases: move low-level mechanism into utils/
Define utils/ as low-level mechanism with no business flow, leaving the root as
the business: index.ts (the flow) and release-plan.ts (the rules). The mechanical
modules — changelog.ts (parsing), github.ts (REST client), git.ts (git/CI
queries) and types.ts — move into utils/. getPackageDirsToSync carried a sliver
of "what to sync" policy, so it's hoisted up into index.ts, leaving git.ts as
pure git wrappers.
