+++ .github/workflows/sync-github-releases/utils/types.ts
